[Fix, Refactor] chat_history.py 생성, generator_main.py make_prompt.py 리팩토링
